Clean up bsdf functions
