+++ src/shader/bsdf/ggx_functions.glsl.js
@@ -1,4 +1,4 @@
-export const ggxGLSL = /* glsl */`
+export const ggx_functions = /* glsl */`
 
 	// The GGX functions provide sampling and distribution information for normals as output so
 	// in order to get probability of scatter direction the half vector must be computed and provided.
