+++ src/shader/bsdf/index.js
@@ -1,0 +1,5 @@
+export * from './bsdf_functions.glsl.js';
+export * from './fog_functions.glsl.js';
+export * from './ggx_functions.glsl.js';
+export * from './iridescence_functions.glsl.js';
+export * from './sheen_functions.glsl.js';
