+++ src/shader/structs/index.js
@@ -2,3 +2,4 @@ export * from './camera_struct.glsl.js';
 export * from './equirect_struct.glsl.js';
 export * from './lights_struct.glsl.js';
 export * from './material_struct.glsl.js';
+export * from './surface_record_struct.glsl.js';
