+++ src/shader/structs/surface_record_struct.glsl.js
@@ -1,0 +1,63 @@
+export const surface_record_struct = /* glsl */`
+
+	struct SurfaceRecord {
+
+		// surface type
+		bool volumeParticle;
+
+		// geometry
+		vec3 faceNormal;
+		bool frontFace;
+		vec3 normal;
+		mat3 normalBasis;
+		mat3 normalInvBasis;
+
+		// cached properties
+		float eta;
+		float f0;
+
+		// material
+		float roughness;
+		float filteredRoughness;
+		float metalness;
+		vec3 color;
+		vec3 emission;
+
+		// transmission
+		float ior;
+		float transmission;
+		bool thinFilm;
+		vec3 attenuationColor;
+		float attenuationDistance;
+
+		// clearcoat
+		vec3 clearcoatNormal;
+		mat3 clearcoatBasis;
+		mat3 clearcoatInvBasis;
+		float clearcoat;
+		float clearcoatRoughness;
+		float filteredClearcoatRoughness;
+
+		// sheen
+		float sheen;
+		vec3 sheenColor;
+		float sheenRoughness;
+
+		// iridescence
+		float iridescence;
+		float iridescenceIor;
+		float iridescenceThickness;
+
+		// specular
+		vec3 specularColor;
+		float specularIntensity;
+	};
+
+	struct ScatterRecord {
+		float specularPdf;
+		float pdf;
+		vec3 direction;
+		vec3 color;
+	};
+
+`;
